+++ WiselySplitServer/src/main/java/ca/mohawkCollege/WiselySplitServer/services/user/UserService.java
@@ -7,12 +7,12 @@ import java.util.Optional;
 
 public interface UserService {
     Optional<User> getUserByEmail(String email);
-    Optional<User> getUserById(int id);
+    Optional<User> getUserById(long id);
     User createUser(User user);
     User updateUser(User user);
-    void deleteUser(int id);
+    void deleteUser(long id);
     List<User> getAllUsers();
-    Map<String, Object> getUserConnections(int userId);
+    Map<String, Object> getUserConnections(long userId);
     public boolean checkUserNameExists(String username);
     public boolean checkEmailExists(String email);
 }
